+++ user-service/src/main/java/it/academy/user_service/dto/PageContent.java
@@ -6,13 +6,13 @@ public class PageContent<T> {
     private int number;
     private int size;
     private int totalPages;
-    private int totalElements;
+    private long totalElements;
     private boolean first;
     private int numberOfElements;
     private boolean last;
     List<T> content;
 
-    public PageContent(int number, int size, int totalPages, int totalElements, boolean first, int numberOfElements, boolean last, List<T> content) {
+    public PageContent(int number, int size, int totalPages, long totalElements, boolean first, int numberOfElements, boolean last, List<T> content) {
         this.number = number;
         this.size = size;
         this.totalPages = totalPages;
@@ -47,7 +47,7 @@ public class PageContent<T> {
         this.totalPages = totalPages;
     }
 
-    public int getTotalElements() {
+    public long getTotalElements() {
         return totalElements;
     }
