+++ user-service/src/main/java/it/academy/user_service/mappers/PageContentMapper.java
@@ -6,11 +6,10 @@ import org.springframework.data.domain.Page;
 
 public class PageContentMapper<T> {
     public PageContent<T> map(Page<T> dtoPage) {
-
         return new PageContent<>(dtoPage.getNumber(),
                 dtoPage.getSize(),
                 dtoPage.getTotalPages(),
-                (int) dtoPage.getTotalElements(),
+                dtoPage.getTotalElements(),
                 dtoPage.isFirst(),
                 dtoPage.getNumberOfElements(),
                 dtoPage.isLast(),
